Feature/default image s3 manager

See merge request teamrot/cv_pipeline_team_rot!66

Former-commit-id: 40c29015d7df9c62b3142b1c955a45e92b2a5a67

+++ sys-src/frontend/src/Components/Connections/awsS3.js
@@ -112,13 +112,19 @@ class S3Manager {
      */
     async deleteImageFromS3(imageKey) {
         // set delete params
-        const params = {
-            Bucket: this.bucketName,
-            Key: imageKey,
-        };
 
-        // delete image from S3
-        return this.S3.deleteObject(params).promise();
+        if (imageKey === "defaultImage.jpg") {
+            return Promise.reject(new Error("Can't delete default image with key: defaultImage.jpg"));
+        }
+        else {
+            const params = {
+                Bucket: this.bucketName,
+                Key: imageKey,
+            };
+    
+            // delete image from S3
+            return this.S3.deleteObject(params).promise();
+        }
     }
 
     /**
@@ -136,16 +142,17 @@ class S3Manager {
             .then( (res) => {
                 // if no objects in bucket
                 if (res.Contents.length === 0) {
-                    return Promise.resolve("No images to delete");
+                    return Promise.reject(new Error("No images to delete"));
                 }
 
                 // get list of all objects
                 const allObjectsFromBucket = res.Contents.map( (object) => ({Key: object.Key, }));
+                const allObjectsFiltered = allObjectsFromBucket.filter( (element) => { return element.Key !== "defaultImage.jpg" });
                 
                 // set delete params
                 const deleteParams = {
                     Bucket: this.bucketName,
-                    Delete: { Objects: allObjectsFromBucket}
+                    Delete: { Objects: allObjectsFiltered}
                 };
 
                 // delete all objects from bucket
